requirements.txt
torch>=1.10.0
torchvision
numpy
opencv-python
timm>=0.4.12
open_clip_torch>=2.0.0
einops